report-job-status: fix branch link in message
The previous logic on building the branch link was incorrect because
the `branch = ref.split("/")[2]` expression did not work as expected
when the branch was named, for example, `someone/feature-1`. So we had
the corrupted link in the message. Moreover, if the ref was a tag, the
link was incorrect as well. Now these two cases are considered.