prepare-checkout: delete stale origin references
This patch adds one more step to the action to delete stale references
associated with `origin`. This will not affect local branches, and it
will not change anything remote, but it will update the local references
to remote branches to avoid errors like this:

    error: cannot lock ref 'refs/remotes/origin/<repo>/<branch>': 'refs/remotes/origin/<repo>' exists; cannot create 'refs/remotes/origin/<repo>/<branch>'
     ! [new branch] <repo>/<branch> -> origin/<repo>/<branch> (unable to update local ref)

However, this can be disabled via the new `prune` option, just set it
to `false`.

Note that submodules are affected as well.